requirements.txt
langchain
langchain_openai
python-dotenv
requests
beautifulsoup4
markdown
unstructured
langchain_community

streamlit
